requirements.txt
odoorpc
workalendar
